test: merge DhisMockMvcControllerTest into ControllerTestBase
ControllerTestBase now extends DhisConvenienceTest. This allows us to
base our web related test on some common integration test base class
also used for service level tests.

* move web specific assertions into an Assertions class in the
  dhis-test-web-api module
