+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/webapi/controller/PdfFormControllerTest.java
@@ -27,6 +27,7 @@
  */
 package org.hisp.dhis.webapi.controller;
 
+import static org.hisp.dhis.webapi.Assertions.assertWebMessage;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
 import org.hisp.dhis.dataset.DataSet;
