+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/webapi/controller/ProgramIndicatorControllerTest.java
@@ -29,6 +29,7 @@ package org.hisp.dhis.webapi.controller;
 
 import static org.hisp.dhis.web.WebClient.Body;
 import static org.hisp.dhis.web.WebClient.ContentType;
+import static org.hisp.dhis.webapi.Assertions.assertWebMessage;
 import static org.springframework.http.MediaType.TEXT_PLAIN_VALUE;
 
 import org.hisp.dhis.web.HttpStatus;
