+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/webapi/controller/ValidationControllerTest.java
@@ -28,6 +28,7 @@
 package org.hisp.dhis.webapi.controller;
 
 import static org.hisp.dhis.utils.Assertions.assertStartsWith;
+import static org.hisp.dhis.webapi.Assertions.assertWebMessage;
 
 import org.hisp.dhis.web.HttpStatus;
 import org.hisp.dhis.webapi.PostgresControllerIntegrationTestBase;
