+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/webapi/controller/userdatastore/UserDatastoreOrderControllerTest.java
@@ -27,6 +27,8 @@
  */
 package org.hisp.dhis.webapi.controller.userdatastore;
 
+import static org.hisp.dhis.webapi.Assertions.assertJson;
+
 import java.util.List;
 import javax.servlet.http.HttpServletResponse;
 import org.hisp.dhis.datastore.DatastoreParams;
